Add holochain prefix to crates
