+++ Cargo.toml
@@ -7,10 +7,10 @@ members = [
   "crates/lib3h_protocol",
   "crates/lib3h",
   "crates/tools/half_busy_chat",
-  "crates/persistence_api",
-  "crates/persistence_mem",
-  "crates/persistence_file",
-  "crates/persistence_pickle",
+  "crates/holochain_persistence_api",
+  "crates/holochain_persistence_mem",
+  "crates/holochain_persistence_file",
+  "crates/holochain_persistence_pickle",
   "crates/holochain_json_api",
-  "crates/json_derive"
+  "crates/holochain_json_derive"
 ]
