+++ crates/holochain_persistence_api/Cargo.toml
@@ -1,14 +1,14 @@
 [package]
-name = "holochain_json_api"
+name = "holochain_persistence_api"
 version = "0.0.1-alpha1"
 authors = ["Holochain Core Dev Team <devcore@holochain.org>"]
 build = "build.rs"
-description = "json features and utilities for holochain"
+description = "lib3h persistence api for content addressable storage and entity attribute value indexes"
 keywords = ["holochain", "holo", "persistence", "cas", "eav"]
 categories = ["database"]
 license = "Apache-2.0"
 readme = "README.md"
-documentation = "https://docs.rs/lib3h_persistence"
+documentation = "https://docs.rs/persistence"
 repository = "https://github.com/holochain/lib3h"
 
 [dependencies]
@@ -30,6 +30,7 @@ hcid = "=0.0.6"
 shrinkwraprs = "=0.2.1"
 rust-base58 = "=0.0.4"
 objekt= "=0.1.2"
+holochain_json_api = { path = "../holochain_json_api" }
 holochain_json_derive = { path = "../holochain_json_derive" }
 uuid = { version = "=0.7.1", features = ["v4"] }
