--- crates/persistence_api/Cargo.toml
@@ -1,38 +1,0 @@
-[package]
-name = "persistence_api"
-version = "0.0.1-alpha1"
-authors = ["Holochain Core Dev Team <devcore@holochain.org>"]
-build = "build.rs"
-description = "lib3h persistence api for content addressable storage and entity attribute value indexes"
-keywords = ["holochain", "holo", "persistence", "cas", "eav"]
-categories = ["database"]
-license = "Apache-2.0"
-readme = "README.md"
-documentation = "https://docs.rs/persistence"
-repository = "https://github.com/holochain/lib3h"
-
-[dependencies]
-arrayref = "=0.3.5"
-base64 = "=0.10.1"
-chrono = "=0.4.6"
-serde = "=1.0.89"
-serde_derive = "=1.0.89"
-serde_json = { version = "=1.0.39", features = ["preserve_order"] }
-multihash = "=0.8.0"
-futures-preview = "=0.3.0-alpha.12"
-futures-core-preview = "=0.3.0-alpha.12"
-futures-channel-preview = "=0.3.0-alpha.12"
-futures-executor-preview = "=0.3.0-alpha.12"
-futures-io-preview = "=0.3.0-alpha.12"
-futures-sink-preview = "=0.3.0-alpha.12"
-futures-util-preview = "=0.3.0-alpha.12"
-hcid = "=0.0.6"
-shrinkwraprs = "=0.2.1"
-rust-base58 = "=0.0.4"
-objekt= "=0.1.2"
-holochain_json_api = { path = "../holochain_json_api" }
-json_derive = { path = "../json_derive" }
-uuid = { version = "=0.7.1", features = ["v4"] }
-
-[dev-dependencies]
-maplit = "=1.0.1"
