--- crates/persistence_file/Cargo.toml
@@ -1,28 +1,0 @@
-[package]
-name = "persistence_file"
-version = "0.0.1-alpha1"
-authors = ["Holochain Core Dev Team <devcore@lib3h.org>"]
-edition = "2018"
-description = "lib3h persistence for content addressable storage and entity attribute value indexes. A file system based implementation using directory hieraarchies to navigate data."
-keywords = ["holochain", "holo", "persistence", "cas", "eav"]
-categories = ["database"]
-license = "Apache-2.0"
-readme = "README.md"
-documentation = "https://docs.rs/persistence_file"
-repository = "https://github.com/holochain/lib3h"
-
-
-[dependencies]
-serde = "=1.0.89"
-serde_derive = "=1.0.89"
-serde_test = "=1.0.89"
-multihash = "=0.8.0"
-persistence_api = { path = "../persistence_api" }
-holochain_json_api = { path = "../holochain_json_api" }
-lazy_static = "=1.2.0"
-glob = "=0.3.0"
-uuid = { version = "=0.7.1", features = ["v4"] }
-
-
-[dev-dependencies]
-tempfile = "=3.0.7"
